create geometry (3D)
